refactor: Expression Test 분리
